backend/requirements.txt
# FastAPI Backend for Physical AI Textbook
fastapi==0.123.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0
openai==1.12.0
qdrant-client==1.9.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-multipart==0.0.18
httpx==0.26.0
tiktoken==0.5.2
tqdm==4.66.3
